Switch push-based deploy to a checkout-per-push model
The problem with doing an archive/untar operation onto an existing directory
(/etc/puppet) is that deleted files never go away.  This doesn't come up often,
but when it does it produces all sorts of hard-to-debug problems.  So we do
this instead.  Sadly, it requires a bit of wrangling to make librarian-puppet
play nicely with the other children and not re-install every module on every
run.

This commit also adds a bunch of other nice features that I've had floating
around for a while.